+++ src/directives/formly-form.test.js
@@ -20,6 +20,18 @@ module.exports = ngModule => {
       expect(el.prop('nodeName').toLowerCase()).to.equal('form');
     });
 
+    it(`should not allow sibling forms to override each other on a parent form`, () => {
+      const el = compileAndDigest(`
+        <form name="parent">
+          <formly-form form="form1"></formly-form>
+          <formly-form form="form2"></formly-form>
+        </form>
+      `);
+      var scope = el.scope();
+      expect(scope.parent).to.have.property('formly_1');
+      expect(scope.parent).to.have.property('formly_2');
+    });
+
     function compileAndDigest(template) {
       const el = $compile(template)(scope);
       scope.$digest();